requirements.txt
streamlit
langchain
openai
faiss-cpu
requests
pickle-mixin
